Added vql functions to manipulate notebooks
This allows for automations around creating custom notebooks.

Also added tests and refactored notebook code:

- Notebooks uploads have been moved to the cells so they can be
  versioned.

- Notebook attachments are not versions and exist in the notebook top
  level.

+++ services/notebook/downloads.go
@@ -3,13 +3,14 @@ package notebook
 import (
 	"context"
 	"errors"
-	"strings"
+	"os"
 	"time"
 
 	api_proto "www.velocidex.com/golang/velociraptor/api/proto"
 	"www.velocidex.com/golang/velociraptor/datastore"
 	"www.velocidex.com/golang/velociraptor/file_store"
 	"www.velocidex.com/golang/velociraptor/file_store/api"
+	"www.velocidex.com/golang/velociraptor/file_store/path_specs"
 	"www.velocidex.com/golang/velociraptor/paths"
 	"www.velocidex.com/golang/velociraptor/reporting"
 )
@@ -35,7 +36,7 @@ func (self *NotebookStoreImpl) GetAvailableDownloadFiles(
 	notebook_id string) (*api_proto.AvailableDownloads, error) {
 
 	download_path := paths.NewNotebookPathManager(notebook_id).
-		HtmlExport().Dir()
+		HtmlExport("X").Dir()
 
 	return reporting.GetAvailableDownloadFiles(self.config_obj, download_path)
 }
@@ -46,30 +47,37 @@ func (self *NotebookStoreImpl) GetAvailableUploadFiles(notebook_id string) (
 
 	notebook_path_manager := paths.NewNotebookPathManager(notebook_id)
 	file_store_factory := file_store.GetFileStore(self.config_obj)
-	files, err := file_store_factory.ListDirectory(
-		notebook_path_manager.AttachmentDirectory())
+
+	notebook, err := self.GetNotebook(notebook_id)
 	if err != nil {
 		return nil, err
 	}
+	for _, cell_metadata := range notebook.CellMetadata {
+		cell_manager := notebook_path_manager.Cell(
+			cell_metadata.CellId, cell_metadata.CurrentVersion)
 
-	for _, item := range files {
-		ps := item.PathSpec()
-		parts := strings.SplitN(ps.Base(), "/", 2)
-		if len(parts) < 2 {
-			continue
+		err := api.Walk(file_store_factory, cell_manager.UploadsDir(),
+			func(ps api.FSPathSpec, info os.FileInfo) error {
+				stat, err := file_store_factory.StatFile(ps)
+				if err != nil {
+					return nil
+				}
+
+				result.Files = append(result.Files, &api_proto.AvailableDownloadFile{
+					Name: ps.Base(),
+					Size: uint64(stat.Size()),
+					Date: stat.ModTime().UTC().Format(time.RFC3339),
+					Type: api.GetExtensionForFilestore(ps),
+					Stats: &api_proto.ContainerStats{
+						Components: ps.Components(),
+					},
+				})
+				return nil
+			})
+		if err != nil {
+			return nil, err
 		}
-
-		result.Files = append(result.Files, &api_proto.AvailableDownloadFile{
-			Name: parts[1],
-			Size: uint64(item.Size()),
-			Date: item.ModTime().UTC().Format(time.RFC3339),
-			Type: api.GetExtensionForFilestore(ps),
-			Stats: &api_proto.ContainerStats{
-				Components: ps.Components(),
-			},
-		})
 	}
-
 	return result, nil
 }
 
@@ -86,8 +94,12 @@ func (self *NotebookStoreImpl) RemoveAttachment(ctx context.Context,
 	}
 
 	notebook_path_manager := paths.NewNotebookPathManager(notebook_id)
-	attachment_path := notebook_path_manager.AttachmentDirectory().
-		AddUnsafeChild(components[len(components)-1])
+	attachment_path := path_specs.NewUnsafeFilestorePath(components...)
+	if !path_specs.IsSubPath(
+		notebook_path_manager.AttachmentDirectory(),
+		attachment_path) {
+		return errors.New("Attachment must be within the notebook directory")
+	}
 
 	file_store_factory := file_store.GetFileStore(self.config_obj)
 	return file_store_factory.Delete(attachment_path)
